- Extracted multistep function from app.js into its own file. - Minor edit to comment in server.js
